+++ test/test_ethernet/test_main.cpp
@@ -1122,6 +1122,36 @@ static void test_server_state() {
                             "Expected default TCP max. listeners");
 }
 
+// Tests setting a server port of zero to have the system choose one.
+static void test_server_zero_port() {
+  TEST_ASSERT_TRUE_MESSAGE(Ethernet.begin(kStaticIP, kSubnetMask, kGateway),
+                           "Expected successful Ethernet start");
+
+  server = std::make_unique<EthernetServer>();
+
+  TEST_ASSERT_FALSE_MESSAGE(static_cast<bool>(*server), "Expected not listening");
+  TEST_ASSERT_EQUAL_MESSAGE(-1, server->port(), "Expected invalid port");
+  TEST_ASSERT_TRUE_MESSAGE(server->begin(0), "Expected TCP listen success");
+  TEST_ASSERT_TRUE_MESSAGE(static_cast<bool>(*server), "Expected listening");
+  int32_t port = server->port();
+  TEST_ASSERT_NOT_EQUAL_MESSAGE(0, port, "Expected non-zero port");
+  TEST_MESSAGE(format("Server port = %" PRId32, port).data());
+  server->end();
+  TEST_ASSERT_FALSE_MESSAGE(static_cast<bool>(*server), "Expected not listening");
+  TEST_ASSERT_EQUAL_MESSAGE(-1, server->port(), "Expected invalid port");
+
+  server = std::make_unique<EthernetServer>(0);
+
+  TEST_ASSERT_FALSE_MESSAGE(static_cast<bool>(*server), "Expected not listening");
+  TEST_ASSERT_EQUAL_MESSAGE(0, server->port(), "Expected zero port");
+  server->begin();
+  TEST_ASSERT_TRUE_MESSAGE(static_cast<bool>(*server), "Expected listening");
+  TEST_ASSERT_NOT_EQUAL_MESSAGE(0, server->port(), "Expected zero port");
+  server->end();
+  TEST_ASSERT_FALSE_MESSAGE(static_cast<bool>(*server), "Expected not listening");
+  TEST_ASSERT_EQUAL_MESSAGE(-1, server->port(), "Expected invalid port");
+}
+
 // Tests state from some of the other classes.
 static void test_other_state() {
   TEST_ASSERT_EQUAL_MESSAGE(DNS_MAX_SERVERS, DNSClient::maxServers(), "Expected default DNS max. servers");
@@ -1227,6 +1257,7 @@ void setup() {
   RUN_TEST(test_client_options);
   RUN_TEST(test_client_diffserv);
   RUN_TEST(test_server_state);
+  RUN_TEST(test_server_zero_port);
   RUN_TEST(test_other_state);
   RUN_TEST(test_raw_frames);
   UNITY_END();